Close #48
I was already fascinated by the ideas proposed by the
chaos engineering discipline, but the possibility of
trying to minimize the problem space using LDFI is
fantastic. I also have found it very insightful to thinking
about fault-tolerance as redundancy.

I need to go more in-depth on the subject, but it
seems promising for the things that I want to explore.